feature: Project structure separated by domain
