+++ abacatepay/base/___init___.py
@@ -1,0 +1,3 @@
+"""
+Module responsible for managing all objects that serve as a foundation for other objects.
+"""
