+++ abacatepay/billings/client.py
@@ -1,14 +1,14 @@
-from ._constants import (
+from ..constants import (
     BASE_URL,
     BILLING_KINDS,
     BILLING_METHODS,
 )
+from ..products import Product
 from .models import (
-    Product,
     BillingResponse,
     Customer,
 )
-from ._base_client import BaseClient
+from ..base.client import BaseClient
 from logging import getLogger
 
 logger = getLogger(__name__)
