+++ abacatepay/products/__init__.py
@@ -1,0 +1,3 @@
+from .models import Product
+
+__all__ = ['Product']
