+++ abacatepay/products/models.py
@@ -1,5 +1,6 @@
 from pydantic import BaseModel
 
+
 class Product(BaseModel):
     externalId: str
     name: str
