+++ tests/integration/test_auth.py
@@ -1,6 +1,6 @@
 import pytest
 from abacatepay import AbacatePay
-from abacatepay.utils._exceptions import UnauthorizedRequest
+from abacatepay.utils.exceptions import UnauthorizedRequest
 
 
 def test_wrong_key_running_function(invalid_token_response):
